Commiting the changes we made together today.  Mostly work on the shoot systm.
i included a field in GameView that initializes the speed of a bullet (the valued we are checking and changing on the counter) so that if we want to later alter the speed of a bullet in game, we have the ability to do so.

I also included a quick kill function to the shoot system so I reduce lag caused by the endless creation of bullets.  We can port this out to a different system later.